[addons][peripheral] move PeripheralUtils.h to own "peripheral" folder
This done to hold the "addon-instance" dir more clean and to have about
every instance type only one file there. If more are needed use them
an extra folder.

SO can more easy seen which instance types are possible.
