+++ xbmc/addons/kodi-dev-kit/include/kodi/addon-instance/CMakeLists.txt
@@ -4,7 +4,6 @@ set(HEADERS AudioDecoder.h
             ImageDecoder.h
             Inputstream.h
             Peripheral.h
-            PeripheralUtils.h
             PVR.h
             Screensaver.h
             VFS.h
