+++ xbmc/addons/kodi-dev-kit/include/kodi/addon-instance/peripheral/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(HEADERS PeripheralUtils.h)
+
+if(NOT ENABLE_STATIC_LIBS)
+  core_add_library(addons_kodi-dev-kit_include_kodi_addon-instance_peripheral)
+endif()
