+++ xbmc/peripherals/addons/PeripheralAddonTranslator.h
@@ -9,7 +9,6 @@
 #pragma once
 
 #include "addons/kodi-dev-kit/include/kodi/addon-instance/Peripheral.h"
-#include "addons/kodi-dev-kit/include/kodi/addon-instance/PeripheralUtils.h"
 #include "input/joysticks/DriverPrimitive.h"
 #include "input/joysticks/JoystickTypes.h"
 #include "input/mouse/MouseTypes.h"
